# backend/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "REST API for Gajian Aman personal finance tracker.\nProvides endpoints for managing accounts and transactions.",
    "title": "Gajian Aman API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "paths": {
    "/api/accounts": {
      "get": {
        "description": "Returns all accounts belonging to the authenticated user, ordered by default status and name.",
        "tags": [
          "accounts"
        ],
        "summary": "List all accounts.",
        "operationId": "listAccounts",
        "responses": {
          "200": {
            "$ref": "#/responses/accountListResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "description": "Creates a new financial account for the authenticated user.\nThe first account created will automatically be set as default.",
        "tags": [
          "accounts"
        ],
        "summary": "Create a new account.",
        "operationId": "createAccount",
        "parameters": [
          {
            "description": "Account creation request",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateAccountDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/accountResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/accounts/check": {
      "get": {
        "description": "Returns whether the authenticated user has at least one account.\nUsed for onboarding flow to redirect users without accounts.",
        "tags": [
          "accounts"
        ],
        "summary": "Check if user has accounts.",
        "operationId": "hasAccounts",
        "responses": {
          "200": {
            "$ref": "#/responses/hasAccountsResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "description": "Returns details of a specific account belonging to the authenticated user.",
        "tags": [
          "accounts"
        ],
        "summary": "Get account by ID.",
        "operationId": "getAccount",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "Account ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/accountResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "description": "Soft-deletes the specified account. Cannot delete the last remaining account.",
        "tags": [
          "accounts"
        ],
        "summary": "Delete an account.",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "Account ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/responses/noContentResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "patch": {
        "description": "Updates the specified account. Only provided fields will be updated.\nSetting is_default to true will clear the default flag from other accounts.",
        "tags": [
          "accounts"
        ],
        "summary": "Update an account.",
        "operationId": "updateAccount",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "Account ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Account update request",
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateAccountDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/accountResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          },
          "401": {
            "$ref": "#/responses/errorResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "AccountDTO": {
      "description": "Account represents a user's financial account",
      "type": "object",
      "required": [
        "name",
        "type"
      ],
      "properties": {
        "balance": {
          "description": "Current balance in the account (in IDR)",
          "type": "number",
          "format": "double",
          "x-go-name": "Balance",
          "example": 1500000
        },
        "id": {
          "description": "The unique identifier of the account",
          "type": "string",
          "x-go-name": "ID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "is_default": {
          "description": "Whether this is the default account for transactions",
          "type": "boolean",
          "x-go-name": "IsDefault",
          "example": true
        },
        "name": {
          "description": "The display name of the account",
          "type": "string",
          "x-go-name": "Name",
          "example": "BCA Savings"
        },
        "type": {
          "description": "Account type: 0=Cash, 1=Bank, 2=E-Wallet, 3=Credit Card, 4=Investment",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Type",
          "example": 1
        },
        "type_name": {
          "description": "Human-readable account type name",
          "type": "string",
          "x-go-name": "TypeName",
          "example": "Bank"
        }
      },
      "x-go-package": "github.com/prgilangdwi/gajianaman/internal/handler"
    },
...
    "CreateAccountDTO": {
      "description": "CreateAccountRequest represents the request body for creating an account",
      "type": "object",
      "required": [
        "name",
        "type"
      ],
      "properties": {
        "balance": {
          "description": "Initial balance in the account (in IDR)",
          "type": "number",
          "format": "double",
          "x-go-name": "Balance",
          "example": 1000000
        },
        "is_default": {
          "description": "Set as default account",
          "type": "boolean",
          "x-go-name": "IsDefault",
          "example": false
        },
        "name": {
          "description": "The display name of the account",
          "type": "string",
          "x-go-name": "Name",
          "example": "BCA Savings"
        },
        "type": {
          "description": "Account type: 0=Cash, 1=Bank, 2=E-Wallet, 3=Credit Card, 4=Investment",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Type",
          "example": 1
        }
      },
      "x-go-package": "github.com/prgilangdwi/gajianaman/internal/handler"
    },
...
    "ErrorResponse": {
      "description": "ErrorResponse is the standard error response envelope",
      "type": "object",
      "properties": {
        "error": {
          "description": "Error code for client handling",
          "type": "string",
          "x-go-name": "Error",
          "example": "INVALID_ACCOUNT_ID"
        },
        "message": {
          "description": "Human-readable error message",
          "type": "string",
          "x-go-name": "Message",
          "example": "Invalid account ID"
        },
        "success": {
          "description": "Whether the request was successful",
          "type": "boolean",
          "x-go-name": "Success",
          "example": false
        }
      },
      "x-go-package": "github.com/prgilangdwi/gajianaman/internal/handler"
    },
    "HasAccountsDTO": {
      "description": "HasAccountsResponse represents the response for checking if user has accounts",
      "type": "object",
      "properties": {
        "has_accounts": {
          "description": "Whether the user has at least one account",
          "type": "boolean",
          "x-go-name": "HasAccounts",
          "example": true
        }
      },
      "x-go-package": "github.com/prgilangdwi/gajianaman/internal/handler"
    },
...
    "UpdateAccountDTO": {
      "description": "UpdateAccountRequest represents the request body for updating an account",
      "type": "object",
      "properties": {
        "is_default": {
          "description": "Set as default account",
          "type": "boolean",
          "x-go-name": "IsDefault",
          "example": true
        },
        "name": {
          "description": "The display name of the account",
          "type": "string",
          "x-go-name": "Name",
          "example": "BCA Savings Updated"
        },
        "type": {
          "description": "Account type: 0=Cash, 1=Bank, 2=E-Wallet, 3=Credit Card, 4=Investment",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Type",
          "example": 1
        }
      },
      "x-go-package": "github.com/prgilangdwi/gajianaman/internal/handler"
    },
...
  },
  "responses": {
    "accountListResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/AccountDTO"
            },
            "x-go-name": "Data"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "success": {
            "type": "boolean",
            "x-go-name": "Success"
          }
        }
      }
    },
    "accountResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/definitions/AccountDTO"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "success": {
            "type": "boolean",
            "x-go-name": "Success"
          }
        }
      }
    },
...
    "errorResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ErrorResponse"
      }
    },
    "hasAccountsResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/definitions/HasAccountsDTO"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "success": {
            "type": "boolean",
            "x-go-name": "Success"
          }
        }
      }
    },
    "noContentResponse": {
      "description": ""
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Supabase JWT token. Format \"Bearer {token}\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}